refactor livehunt module and http router
