+++ src/livehunt/mod.rs
@@ -1,0 +1,9 @@
+mod client;
+mod notifications;
+mod types;
+
+pub use client::*;
+pub use types::*;
+
+#[cfg(test)]
+mod tests;
