+++ src/tests/mod.rs
@@ -36,12 +36,3 @@ mod integration_tests;
 mod additional_coverage_tests;
 #[cfg(test)]
 mod async_coverage_tests;
-// #[cfg(test)]
-// mod mcp_comprehensive_tests;
-// #[cfg(test)]
-// mod files_comprehensive_tests;
-// #[cfg(test)]
-// mod client_comprehensive_tests;
-// #[cfg(test)]
-// mod error_comprehensive_tests;
-// mod api_modules_comprehensive_tests; // Temporarily disabled due to API structure issues
